Move collaboration model.js to collaborationUtils.js
